+++ include/structs.h
@@ -18,7 +18,7 @@ struct Player {
   double x;
   double y;
   double angle;
-  int height;
+  double height;
 };
 
 struct Engine {
